Separate component to view the plot
